+++ Vector3.h
@@ -1,14 +1,14 @@
 #pragma once
-#include "Header.h"
+#include "resources.h"
 class Vector3
 {
 private:
-	DOUBLE length;
+	LONG length;
 public:
-	DOUBLE x, y, z;
+	LONG x, y, z;
 
 	Vector3();
-	Vector3(DOUBLE _x, DOUBLE _y, DOUBLE _z);
+	Vector3(long _x, long _y, long _z);
 	DOUBLE Len();
 	Vector3 Normalized();
 	DOUBLE cos(Vector3 _v3);
